Added a preview tab on all widgets
